[REFACTOR] 리뷰 반영
