+++ user-api/src/main/java/com/biengual/userapi/category/repository/CategoryCustomRepository.java
@@ -1,32 +1,33 @@
 package com.biengual.userapi.category.repository;
 
-import com.biengual.userapi.category.domain.CategoryInfo;
+import java.util.List;
 
-import com.biengual.userapi.category.domain.entity.QCategoryEntity;
-import com.querydsl.core.types.Projections;
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
+import com.biengual.userapi.category.domain.CategoryInfo;
+import com.biengual.userapi.category.domain.QCategoryEntity;
+import com.querydsl.core.types.Projections;
+import com.querydsl.jpa.impl.JPAQueryFactory;
+
+import lombok.RequiredArgsConstructor;
 
 @Repository
 @RequiredArgsConstructor
 public class CategoryCustomRepository {
-    private final JPAQueryFactory queryFactory;
+	private final JPAQueryFactory queryFactory;
 
-    // 서비스에 등록된 모든 카테고리를 조회하기 위한 쿼리
-    public List<CategoryInfo.Category> findAllCategories() {
-        QCategoryEntity categoryEntity = QCategoryEntity.categoryEntity;
+	// 서비스에 등록된 모든 카테고리를 조회하기 위한 쿼리
+	public List<CategoryInfo.Category> findAllCategories() {
+		QCategoryEntity categoryEntity = QCategoryEntity.categoryEntity;
 
-        return queryFactory.select(
-                Projections.constructor(
-                    CategoryInfo.Category.class,
-                    categoryEntity.id,
-                    categoryEntity.name
-                )
-            )
-            .from(categoryEntity)
-            .fetch();
-    }
+		return queryFactory.select(
+				Projections.constructor(
+					CategoryInfo.Category.class,
+					categoryEntity.id,
+					categoryEntity.name
+				)
+			)
+			.from(categoryEntity)
+			.fetch();
+	}
 }
