+++ user-api/src/main/java/com/biengual/userapi/content/presentation/ContentApiController.java
@@ -62,7 +62,7 @@ public class ContentApiController {
 	/**
 	 * 컨텐츠 수정
 	 */
-	@PutMapping("/modify/{id}")
+	@PutMapping("/modify/{contentId}")
 	@Operation(summary = "어드민 - 컨텐츠 수정", description = "어드민 회원이 컨텐츠를 수정합니다.")
 	@ApiResponses(value = {
 		@ApiResponse(responseCode = "200", description = "컨텐츠 수정 성공",
@@ -74,11 +74,11 @@ public class ContentApiController {
 	})
 	public ResponseEntity<Object> modifyContent(
 		@PathVariable
-		Long id,
+		Long contentId,
 		@RequestBody
 		ContentRequestDto.UpdateReq request
 	) {
-		ContentCommand.Modify command = contentDtoMapper.doModify(id, request);
+		ContentCommand.Modify command = contentDtoMapper.doModify(contentId, request);
 		contentFacade.modifyContent(command);
 
 		return ResponseEntityFactory.toResponseEntity(CONTENT_MODIFY_SUCCESS);
@@ -87,7 +87,7 @@ public class ContentApiController {
 	/**
 	 * 컨텐츠 비활성화
 	 */
-	@PutMapping("/deactivate/{id}")
+	@PutMapping("/deactivate/{contentId}")
 	@Operation(summary = "어드민 - 컨텐츠 비활성화", description = "어드민 회원이 컨텐츠를 비활성화합니다.")
 	@ApiResponses(value = {
 		@ApiResponse(responseCode = "200", description = "컨텐츠 비활성화 성공",
@@ -99,9 +99,9 @@ public class ContentApiController {
 	})
 	public ResponseEntity<Object> deactivateContent(
 		@PathVariable
-		Long id
+		Long contentId
 	) {
-		contentFacade.deactivateContent(id);
+		contentFacade.deactivateContent(contentId);
 
 		return ResponseEntityFactory.toResponseEntity(CONTENT_DEACTIVATE_SUCCESS);
 	}
